shoutcast/requirements/project.txt
# project.txt is a pip requirements file which describes the distributions
# required by your project to run.

--requirement=base.txt

# Put project-specific requirements here.
# See http://pip-installer.org/requirement-format.html for more information.


mutagen
pyechonest
slumber
django-tastypie
lxml
django-celery
redis
whoosh
django-haystack
